refactor: move two interface back to entity package
