+++ src/main/kotlin/com/wafflestudio/csereal/common/entity/AttachmentAttachable.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.csereal.common.domain
+package com.wafflestudio.csereal.common.entity
 
 import com.wafflestudio.csereal.core.resource.attachment.database.AttachmentEntity
