+++ src/main/kotlin/com/wafflestudio/csereal/core/council/database/CouncilEntity.kt
@@ -1,7 +1,7 @@
 package com.wafflestudio.csereal.core.council.database
 
 import com.wafflestudio.csereal.common.entity.BaseTimeEntity
-import com.wafflestudio.csereal.common.domain.MainImageAttachable
+import com.wafflestudio.csereal.common.entity.MainImageAttachable
 import com.wafflestudio.csereal.core.council.dto.ReportCreateRequest
 import com.wafflestudio.csereal.core.resource.mainImage.database.MainImageEntity
 import jakarta.persistence.*
