+++ src/main/kotlin/com/wafflestudio/csereal/core/council/database/CouncilFileEntity.kt
@@ -1,7 +1,7 @@
 package com.wafflestudio.csereal.core.council.database
 
 import com.wafflestudio.csereal.common.entity.BaseTimeEntity
-import com.wafflestudio.csereal.common.domain.AttachmentAttachable
+import com.wafflestudio.csereal.common.entity.AttachmentAttachable
 import com.wafflestudio.csereal.core.council.type.CouncilFileType
 import com.wafflestudio.csereal.core.resource.attachment.database.AttachmentEntity
 import jakarta.persistence.*
